requirements.txt
pandas 
google
google-generativeai 
python-dotenv
packaging
google-genai
statistics
transformers
huggingface